# use bigger buffer size when reading and writing large files