Increases the buffer size when reading and writing large files (eg. the conversion into `.xyz.bin` files and reading of the same) from the default 8KiB to 1MB. This means fewer syscalls to the operating system and thus a small speedup in processing.

Here are some comparison tables of the number and time spent in syscalls for the test `laz` file on my computer (captured usince `strace -c`):

**Before**
```
[2025-02-03T11:57:58Z ThreadId(1) DEBUG pullauta::util] [timing: process_tile] Stopping timing. Total: 52.104s elapsed.
% time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ------------------
 68.72    1.412254           1    709497           read
 11.68    0.240137        2638        91           openat
  7.59    0.156014           1    146707           write
  6.52    0.133980        5825        23           munmap
  3.90    0.080175         881        91           close
  1.20    0.024570           8      2762           brk
  0.36    0.007484        3742         2           copy_file_range
  0.01    0.000229           3        74           ioctl
  0.00    0.000099           1        78           mmap
  0.00    0.000053           1        27         7 statx
  0.00    0.000032           3         9           lseek
  0.00    0.000016           3         5           rt_sigaction
  0.00    0.000013           6         2         2 mkdir
  0.00    0.000012           3         4           getrandom
  0.00    0.000011           3         3           sigaltstack
  0.00    0.000007           0        15           mprotect
  0.00    0.000006           6         1           sched_getaffinity
  0.00    0.000005           5         1           fchmod
  0.00    0.000000           0        14           fstat
  0.00    0.000000           0         1           poll
  0.00    0.000000           0         2           pread64
  0.00    0.000000           0         1         1 access
  0.00    0.000000           0         1           execve
  0.00    0.000000           0         1           arch_prctl
  0.00    0.000000           0         1           futex
  0.00    0.000000           0         1           set_tid_address
  0.00    0.000000           0         1           set_robust_list
  0.00    0.000000           0         2           prlimit64
  0.00    0.000000           0         1           rseq
------ ----------- ----------- --------- --------- ------------------
100.00    2.055097           2    859418        10 total
```


**After**
```
[2025-02-03T11:59:28Z ThreadId(1) DEBUG pullauta::util] [timing: process_tile] Stopping timing. Total: 39.818s elapsed.
% time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ------------------
 28.45    0.404708          10     36950           write
 20.83    0.296395          10     28456           read
 19.20    0.273159        3001        91           close
 17.88    0.254447        2796        91           openat
 10.67    0.151796        6324        24           munmap
  2.36    0.033513          12      2765           brk
  0.54    0.007649        3824         2           copy_file_range
  0.04    0.000593           7        79           mmap
  0.01    0.000184           2        74           ioctl
  0.01    0.000126           8        15           mprotect
  0.00    0.000049           3        14           fstat
  0.00    0.000044           1        27         7 statx
  0.00    0.000021          10         2           pread64
  0.00    0.000017           1         9           lseek
  0.00    0.000012           3         4           getrandom
  0.00    0.000009           4         2         2 mkdir
  0.00    0.000008           4         2           prlimit64
  0.00    0.000005           5         1           fchmod
  0.00    0.000005           5         1           futex
  0.00    0.000004           4         1           arch_prctl
  0.00    0.000004           4         1           set_tid_address
  0.00    0.000004           4         1           set_robust_list
  0.00    0.000004           4         1           rseq
  0.00    0.000001           0         3           sigaltstack
  0.00    0.000000           0         1           poll
  0.00    0.000000           0         5           rt_sigaction
  0.00    0.000000           0         1         1 access
  0.00    0.000000           0         1           execve
  0.00    0.000000           0         1           sched_getaffinity
------ ----------- ----------- --------- --------- ------------------
100.00    1.422757          20     68625        10 total
```

As you can see, on my device the total processing time decreases from 52 to 39 seconds, and the number of (relatively expensive) syscalls have decreased :rocket: Performance improvements are of course quite hardware-dependent, but this should be an improvement overall IMO.